+++ src/cqindustry/__init__.py
@@ -1,4 +1,3 @@
-from .Base import Base
 from .Platform import Platform
 from .Ring import Ring
 from .Walkway import Walkway